% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_agrupado}
\alias{da_agrupado}
\title{da_agrupado}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 253 rows and 7 columns.
}
\usage{
da_agrupado
}
\description{
da_agrupado
}
\keyword{datasets}
