% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_metadados}
\alias{da_metadados}
\title{da_metadados}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 271 rows and 6 columns.
}
\usage{
da_metadados
}
\description{
da_metadados
}
\keyword{datasets}
