% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_spss}
\alias{da_spss}
\title{da_spss}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 252 rows and 13 columns.
}
\usage{
da_spss
}
\description{
da_spss
}
\keyword{datasets}
