% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_tidy}
\alias{da_tidy}
\title{da_tidy}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2522293 rows and 9 columns.
}
\usage{
da_tidy
}
\description{
da_tidy
}
\keyword{datasets}
