% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{filtra_nomes}
\alias{filtra_nomes}
\title{Filtra nomes de bebês que fizeram condição contingente para dado grupo e dia}
\usage{
filtra_nomes(da, gr, d)
}
\arguments{
\item{da}{dados \link{da_tidy}}

\item{gr}{grupo (b1, b2, b3)}

\item{d}{1, 2 ou 3}
}
\description{
Filtra nomes de bebês que fizeram condição contingente para dado grupo e dia
}
