% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{hc_grupo}
\alias{hc_grupo}
\title{Gráfico de pressão média por condição basal e contingente}
\usage{
hc_grupo(g)
}
\arguments{
\item{g}{grupo (b1, b2, b3)}
}
\description{
Gráfico de pressão média por condição basal e contingente
}
