% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{p_video_bebe}
\alias{p_video_bebe}
\title{Gráfico de pressao ao longo do tempo}
\usage{
p_video_bebe(da, bebe, d, gr)
}
\arguments{
\item{da}{dados \link{da_tidy}}

\item{bebe}{nome do bebe}

\item{d}{dia}

\item{gr}{grupo (b1, b2, b3)}
}
\description{
Gráfico de pressao ao longo do tempo
}
